@@ -15,21 +15,22 @@ classifiers=
     Intended Audience :: Science/Research
     License :: OSI Approved :: BSD License
     Operating System :: OS Independent
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
     Topic :: Scientific/Engineering
 [options]
 packages = find:
 include_package_data = True
 install_requires =
-        matplotlib>=3.4
-        numpy>=1.21
-        scipy>=1.7
-        pandas>=1.3
-        pathos>=0.2
-python_requires = >=3.8
+        matplotlib
+        numpy
+        scipy
+        pandas
+        pathos
+        tqdm
+        pyarrow
+        shapely
+        opencv
+python_requires = >=3.12
 [options.packages.find]
 where = .
+exclude =
+    telemetry*
